requirements.txt
requests
firebase-admin
google-api-python-client
serpapi
beautifulsoup4
lxml
playwright
playwright-stealth
statistics
